app/protobufs/account/requirements.txt
connexion[swagger-ui]~=2.12.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
pymongo[srv] >= 4.0
prometheus_client
flask
pyopenssl

grpcio >= 1.30.0
grpcio-tools >= 1.30.0
grpc-interceptor >= 0.13.0
